+++ packages/navigator/src/browser/navigator-contribution.ts
@@ -31,6 +31,7 @@ import { FileNavigatorFilter } from './navigator-filter';
 import { WorkspaceNode } from './navigator-tree';
 import { NavigatorContextKeyService } from './navigator-context-key-service';
 import { TabBarToolbarContribution, TabBarToolbarRegistry } from '@theia/core/lib/browser/shell/tab-bar-toolbar';
+import { FileSystemCommands } from '@theia/filesystem/lib/browser/filesystem-frontend-contribution';
 
 export namespace FileNavigatorCommands {
     export const REVEAL_IN_NAVIGATOR: Command = {
@@ -195,7 +196,7 @@ export class FileNavigatorContribution extends AbstractViewContribution<FileNavi
 
         const downloadUploadMenu = [...NAVIGATOR_CONTEXT_MENU, '6_downloadupload'];
         registry.registerMenuAction(downloadUploadMenu, {
-            commandId: FileDownloadCommands.UPLOAD.id,
+            commandId: FileSystemCommands.UPLOAD.id,
             order: 'a'
         });
         registry.registerMenuAction(downloadUploadMenu, {